// src/app/generated/model/supplierCreateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SupplierCreateRequest { 
    /**
     * Name of the supplier
     */
    name: string;
    /**
     * Contact details for the supplier, e.g. phone, email, or address
     */
    contactInfo?: string;
}