// src/app/generated/model/supplierResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SupplierResponse { 
    /**
     * Unique identifier of the supplier
     */
    id?: number;
    /**
     * Name of the supplier
     */
    name?: string;
    /**
     * Contact details for the supplier, e.g. phone, email, or address
     */
    contactInfo?: string;
    /**
     * Whether the supplier is currently active
     */
    active?: boolean;
    /**
     * Timestamp when the supplier was created
     */
    createdAt?: string;
    /**
     * Timestamp when the supplier was last updated
     */
    updatedAt?: string;
}